Update Quiz components to record question results during a quiz
Results are maintained atleast locally to the QuizView with the intent
that they will be pushed to the redux store in a later commit for the
specific deck that the quiz occurred on.

Changes here also includes an Alert window that pops up when the user
navigates back from the quiz notifying the user if they've scored all
the cards yet and the totals if they have scored all cards.